Display version number and date in documentation.
Dude, what the hell…
As much as I like LaTeX for the visual quality which can be achieved
with it, doctex can suck horribly.
E.g. for correctly showing the file version and date in the generated
documentation, you have to put all declarations in the exact same order
as in the tutorial. In other places (the infamous macrocode
environment), whitespace matters, which is highly unusual for TeX, and
only becomes apparent when LaTeX starts spewing errors, and you swear
and curse and cry and read the docs.